# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AURA Dealership Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dashboard/sentiment:
    get:
      operationId: getSentimentAnalysis
      tags: [dashboard]
      summary: AI sentiment analysis over recent lead notes and conversations
      parameters:
        - name: refresh
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: Sentiment analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SentimentAnalysis"
        "502":
          description: Sentiment engine unavailable
...
components:
  schemas:
...
    SentimentDistribution:
      type: object
      required: [positive, neutral, negative]
      properties:
        positive: { type: number }
        neutral: { type: number }
        negative: { type: number }
    SentimentTheme:
      type: object
      required: [theme, sentiment, mentions]
      properties:
        theme: { type: string }
        sentiment: { type: string, enum: [positive, neutral, negative] }
        mentions: { type: integer }
    SentimentHighlight:
      type: object
      required: [leadId, leadName, sentiment, snippet]
      properties:
        leadId: { type: ["integer", "null"] }
        leadName: { type: string }
        sentiment: { type: string, enum: [positive, neutral, negative] }
        snippet: { type: string }
    SentimentAnalysis:
      type: object
      required: [overallScore, overallLabel, summary, distribution, themes, highlights, sampleSize, generatedAt]
      properties:
        overallScore: { type: number }
        overallLabel: { type: string, enum: [positive, neutral, negative] }
        summary: { type: string }
        distribution: { $ref: "#/components/schemas/SentimentDistribution" }
        themes:
          type: array
          items: { $ref: "#/components/schemas/SentimentTheme" }
        highlights:
          type: array
          items: { $ref: "#/components/schemas/SentimentHighlight" }
        sampleSize: { type: integer }
        generatedAt: { type: string, format: date-time }
...